Agents/requirements.txt
# Core dependencies
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
PyGithub>=2.1.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0

# Utility dependencies
pyyaml>=6.0.0
click>=8.1.0
rich>=13.0.0
typing-extensions>=4.8.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0
pre-commit>=3.5.0

# Optional dependencies for enhanced features
httpx>=0.25.0
tenacity>=8.2.0
structlog>=23.2.0
